pyproject.toml
[project]
name = "macrocast"
version = "0.1.0"
description = "Decomposing ML Forecast Gains in Macroeconomic Forecasting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Chan"},
]
keywords = ["forecasting", "macroeconomics", "machine-learning", "FRED"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "statsmodels>=0.14",
    "scipy>=1.11",
    "requests>=2.31",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "tqdm>=4.65",
]

[project.urls]
Documentation = "https://macrocast.github.io/macrocast"
Repository = "https://github.com/macrocast/macrocast"
Issues = "https://github.com/macrocast/macrocast/issues"

[project.optional-dependencies]
ml = [
    "lightgbm>=4.0",
    "torch>=2.0",
    "xgboost>=2.0",
    "pyarrow>=14.0",
]
viz = [
    "matplotlib>=3.7",
    "seaborn>=0.13",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.5",
    "pre-commit>=3.4",
]
docs = [
    "mkdocs-material>=9.4",
    "mkdocstrings[python]>=0.23",
]
rag = [
    "mcp[cli]>=1.2.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pymupdf>=1.23.0",
    "beautifulsoup4>=4.12.0",
    "einops>=0.8.2",
]
all = ["macrocast[ml,viz,dev,docs,rag]"]

[project.scripts]
macrocast = "macrocast.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["macrocast"]

[tool.hatch.build]
include = [
    "macrocast/**/*.py",
    "macrocast/data/specs/*.json",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = [
    "E501",    # line too long — handled by formatter
    "N803",    # argument name should be lowercase — math conventions (X, W)
    "N806",    # variable should be lowercase — math conventions (X_mat, XtX, W, T)
    "N802",    # function name should be lowercase — KFOLD factory is intentional
    "N815",    # mixedCase class attribute — standardize_X/Z are intentional
    "B905",    # zip() without strict= — not needed for internal zip calls
    "SIM103",  # return condition directly — readability preference
]

[tool.ruff.lint.isort]
known-first-party = ["macrocast"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -m 'not network'"
markers = [
    "network: marks tests that require network access (deselect with '-m not network')",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
